requirements.txt
telethon==1.36.0
supabase>=2.11.0
openai>=1.59.0
python-dotenv==1.0.1
apscheduler==3.10.4
python-docx>=1.1.0
requests>=2.31.0
beautifulsoup4>=4.12.0
